refactorizacion de los controladores y rutas se le quito responsabilidad al controlador del dashboard, para crear el controlador de usuarios y de historial, al hacer esto tambien se tuvo que modificar las rutas y se aprovecho de hacer una optimizacion de las mismas
